tests/wpt/tests/css/css-values/calc-size/animation/calc-size-width-interpolation.tentative.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>width: calc-size() animations</title>
<link rel="help" href="https://drafts.csswg.org/css-values-5/#calc-size">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../support/interpolation-testcommon.js"></script>

<style>
.parent {
  display: block;
  width: 200px;
}
.target {
  display: block;
}
.target::before {
  display: block;
  content: "";
  width: 100px;
}
</style>

<body>

<script>
  test_interpolation({
    property: 'width',
    from: 'calc-size(auto, size)',
    to: 'calc-size(auto, size * 2)',
  }, [
    { at: -0.25, expect: '150px' },
    { at: 0, expect: '200px' },
    { at: 0.25, expect: '250px' },
    { at: 0.5, expect: '300px' },
    { at: 0.75, expect: '350px' },
    { at: 1, expect: '400px' },
    { at: 1.25, expect: '450px' },
  ]);

  test_interpolation({
    property: 'width',
    from: neutralKeyframe,
    to: 'calc-size(auto, size * 2)',
  }, [
    { at: -0.25, expect: '150px' },
    { at: 0, expect: '200px' },
    { at: 0.25, expect: '250px' },
    { at: 0.5, expect: '300px' },
    { at: 0.75, expect: '350px' },
    { at: 1, expect: '400px' },
    { at: 1.25, expect: '450px' },
  ]);

  test_interpolation({
    property: 'width',
    from: 'calc-size(min-content, 0px)',
    to: 'calc-size(min-content, size)',
  }, [
    { at: -0.25, expect: '0' },
    { at: 0, expect: '0' },
    { at: 0.25, expect: '25px' },
    { at: 0.5, expect: '50px' },
    { at: 0.75, expect: '75px' },
    { at: 1, expect: '100px' },
    { at: 1.25, expect: '125px' },
  ]);

</script>
